fix compaction-test on ext3/ext4
Empty directories on ext3 and ext4 report a disk space consumption of 4096
bytes, but on xfs (where kudu-gerrit runs) they're just a few bytes. So
when this test was run on a laptop (typically with ext4), the compaction
did reclaim some disk space, but it also generated a whole bunch of new
directories, and the latter outweighed the former.

There's an easy solution to this: exclude directory space usage from the
test's calculation.

Note that this still fails when run with the log block manager; the punched
holes are <4096 bytes in size and thus no on-disk blocks are actually
reclaimed (some portions of blocks are merely zeroed).

Change-Id: Ib7d8a81aaa02e3aaca6905f9d4c853b6fae405b4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6028
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>