LogGC shouldn't rely on the TabletPeer's consensus ready latch
I made a wrong fix to TabletPeer::RunLogGC() by making it wait on consensus to be ready and we
noticed that on a cluster where we restart nodes and don't churn leader elections that we hit
cases where that consensus latch is never changed. This patch simply removes the check.

I tried removing the latch completely but the master currently relies on reaching a steady state
when starting. In particular, CatalogManager::Init() expects that after running
sys_catalog_->WaitUntilRunning() that if it is the leader then that the rest of the insfrastructure
agrees, but that might be wrong and ElectedAsLeaderCb() never fires.
MasterReplicationTest.TestCycleThroughAllMasters hangs in that case.

Change-Id: I3948207a896953bb00e192b7280018b4de5efef3
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6032
Tested-by: jenkins <cloudera-hudson@gerrit.cloudera.org>
Reviewed-by: Todd Lipcon <todd@cloudera.com>