Mvcc: Account for single transactions in MayHaveUncommittedTransactionsAtOrBefore()
Let 5, 10 be the timestamps of two in-flight non commit-wait transactions. When 5 commits
we advance the all_committed_before to 10. But when 10 commits we can't advance it as
there are no other timestamps to advance it to. Meaning all_committed_before == 10 but
10 is also in the committed set.

In this case MayHaveUncommittedTransactionsAtOrBefore(10) should return false, since
we know that all trasactions wil lower timestamps are committed and 10 is itself committed.
However because we were only looking all_committed_before we would do the wrong thing.

This changes that method so that if all_committed_before is precisely equal to the
timestamp the caller is asking about, it also looks on the committed set.

Note that this additional condition should be used rarely and only when the committed
set has a single entry, so it shouldn't represent any meaninful overhead.

Change-Id: I42741315356805353136e42b09251b97f4abdd06
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6036
Tested-by: jenkins <cloudera-hudson@gerrit.cloudera.org>
Reviewed-by: Todd Lipcon <todd@cloudera.com>