Mvcc: Allow to abort an mvcc transaction
Right now even when transactions fail we call MvccManager::CommitTransaction().
This wasn't very problematic before, if transactions had made no changes to a tablet.
However even failed transactions would advance the safe time on the leader.

With transaction aborts and timestamp propagation we need aborted not to advance the
safe time, since a new leader might force a previous leader to start transactions
in the past, as read by the previous leader's clock.

This patch adds an AbortTransaction() method to MvccManager and a test that
makes sure it works. A follow up patch integrates this with TransactionDriver.

Change-Id: I1e08fbea480c76bf702ea31ac97c8685aa342488
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6042
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins <cloudera-hudson@gerrit.cloudera.org>