Mvcc: Advance the safe time with committed transactions, not the clock
We're currently facing the following problem with the logical clock:
- Node A, leader, commits txn 10 and advances safe time to clock_->Now() = 20 (this
  is possible as the clock's logical time may increase arbitrarily based on the number
  of calls to clock_->Now()).
- Node A loses leadership.
- Node B becomes leader and sends txn 11 to Node A, but A can't handle it as it
  thinks all transactions < 20 are already committed. We crash.

Note we could also observe this with the hybrid clock:
- Node A starts applying committed txn 10, which takes a while.
- Node A loses leadership.
- Node B becomes leader and replicates and commits txn 15.
- Node A finishes applying and commits txn 10. Since it took a while, the clock
  now reads 20 to which it update the safe time.
- Node B sends txn 15 to Node A, but A can't handle as it thinks that transactions
  < 20 are already committed.

This patch first splits the concepts of safe and clean timestamps, which
we were using kind of interchangeably before:
- A clean timestamp is a safe timestamp for which all transactions are
  committed. This is the timestamp required to perform repeatable reads.
- A safe timestamp is one for which no transaction with a lower timestamp
  will be started, i.e. all the transactions with a lower timestamp are
  either committed or in-flight.

On the leader the safe timestamp is updated each time a transaction commits.
This is correct since all transactions commit with a timestamp < clock_->Now(),
mandatorily.
On replicas we advance the safe timestamp based on what the leader dictates
it should be. This last part is to be done in a follow up patch.

The main difference vs. the previous impl. is that we now set the safe time
to be the max of all observed committed timestamps (i.e. we store it even if
we can't advance the clean timestamp right away). We then advance the clean
timestamp to either the earliest in flight or to the safe timestamp (we
previously advanced it it to the earliest timestamp or to clock_->Now()).

Change-Id: Ie6ff6e11fe2f435c8974ce8d26af9ed6d6caea59
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6041
Tested-by: jenkins <cloudera-hudson@gerrit.cloudera.org>
Reviewed-by: Todd Lipcon <todd@cloudera.com>