tests: ensure tablet peers are destroyed before tablet servers
In an upcoming commit, blocks will start accessing block manager state at
block destruction time. To do this safely we must ensure that the block
manager is still alive when the blocks are destroyed, which means the
rowsets in tablet peers must not be destroyed after their owning TS.
This commit fixes the few tests that violate this contract.

Additionally, in TabletServerTest, we must ensure that the tablet peer owns
the last reference to each rowset, otherwise rowset destruction is deferred
to the end of the test, after TS destruction in ShutdownAndRebuildTablet().

Change-Id: Ic1b4fab486cec4089755688d07108e9ff0d461e5
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6053
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>