thread: add synchronization-ignoring annotations
I have a block manager unit test that should trigger a data race in TSAN,
yet doesn't. Turns out it's because TSAN interprets some synchronization
performed by a group of threads in SuperviseThread() as communication
between them (i.e. "happens before" relationships). For example, both
AddThread() and RemoveThread() surround the thread's functor and acquire
a mutex. TSAN sees this and establishes an ordering between the threads
instead of looking for races.

Let's add TSAN annotations[1] to tell TSAN to ignore this synchronization.
They appear to work in both clang 3.4 and 3.5. Of course, this also means
telling TSAN not to complain about races on memory accessed inside that
synchronization.

1. http://lists.cs.uiuc.edu/pipermail/llvm-commits/Week-of-Mon-20131007/190668.html

Change-Id: I825ef521099e7c79661c496a9a5abcf7f493693e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6054
Tested-by: jenkins <cloudera-hudson@gerrit.cloudera.org>
Reviewed-by: Todd Lipcon <todd@cloudera.com>