Rename AtomicInt<X>::CompareAndSwap to CompareAndSet, CompareAndSwapVal to CompareAndSwap
In reviewing a recent patch, I ran into some confusion at a call site that was
using AtomicBool::CompareAndSwap(). Currently, that function returns true/false
to indicate success, whereas the base::subtle::*CompareAndSwap functions all
return the _previous_ value (i.e "swap" semantics). In the case of an
AtomicBool, this is particularly confusing, because the type is a bool either
way, but even in the case of integers this can be confusing because the int
return values can implicitly cast to bool.

This changes the naming of these functions to be more consistent with
base::subtle.  In particular, CompareAndSwap() is now called CompareAndSet(),
and returns a bool to indicate success/failure, like the Java API of the same
name.  The old CompareAndSwapVal() function is now called CompareAndSwap()
because it in fact does a swap.

Change-Id: I04adc38246c9d79fcaf259950f1f840ae5d4be5f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6064
Reviewed-by: Andrew Wang <andrew.wang@cloudera.com>
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>