KUDU-372: Maintenance manager should schedule delta compactions
This patch adds a new MaintenanceOp in the tablet to run minor delta compactions. Like its cousin
the delta flush op, there's only one instance per tablet and it picks a RowSet that provides the
best performance improvement.

It has a simple test for the computation of the perf improvement, it also augments mt-tablet-test
to start running delta flushes and minor compactions, and it was also tested on a single node with
a 100% update YCSB workload.

Change-Id: If495eb43c58dbfc081680a0c049e734c8774ceb0
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6072
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>