KUDU-618. Fix deadlock between PeerMessageQueue and Log
This fixes the following deadlock:

- PeerMessageQueue::AppendOperations() holds PeerMessageQueue::queue_lock_
  while waiting to call Log::Append. This blocks because the log's queue
  is full with other in-flight appends.
- Log::AppendThread is trying to call PeerMessageQueue::LocalPeerAppendFinished()
  which blocks trying to acquire PeerMessageQueue::queue_lock_. Because the
  AppendThread is blocked, it is not dequeuing entries and thus never frees
  space in the log.

This fixes the issue by changing PeerMessageQueue to not hold the lock while
it calls down into the log. This is safe so long as we disallow concurrent
appenders on PeerMessageQueue. That's currently the case because the consensus
implementation itself already synchronizes its calls into the queue. I added a
fake mutex to ensure that this doesn't change in the future.

Unfortunately we haven't figured out a way to trigger this in a functional test,
but it seems to reproduce easily in the YCSB setup. I will ask JD to test this
via YCSB before we commit.

Change-Id: I335c945d0d00ba6ece8541cddbd78bb0bb116778
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6073
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins