Bug Fix: Only advance safe time after a trasaction is committed in consensus
We're currently advancing safe time when we prepare a transaction. This is
wrong since a prepared transaction can still be aborted and there is no
guarantee that an overwriting or later transaction from another leader
won't have an earlier timestmap. This would trigger failures in
raft_consensus-itest on master.

The fix is to only advance safe time once we know that the transaction
is consensus committed. This patch does that in the transaction driver.

Change-Id: Ic49d50e7cfd14b1ac2d39e031e283674b756c7c4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6074
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>