KUDU-626: Skip uninitialized OpIds when computing the earliest needed log index
The transactions that are causing problems in KUDU-626 have an OpId that is not initialized.
It's safe to skip them since they'll eventually have an OpId that's greater than what we
already computed in TabletPeer::GetEarliestNeededLogIndex.

Tested that it fixes the issue the same way it was triggered in the Jira above.

Change-Id: I2075201566110749b5ad9f9f06fa0d73e1cebcef
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6078
Reviewed-by: David Alves <david.alves@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins