Bug Fix: On commit, also coalesce clean time for offline transactions
Previously we weren't coalescing the clean time on commit, for offline
transactions. We were doing it only on OfflineAdjustSafeTime() which,
in turn, would only happen when writes were happening.

This was problematic because even though a snapshot was reported as
clean, we would not coalesce and thus MayHaveCommittedTransactionsAtOrAfter()
would return true.

Change-Id: I8b0dd3d9b36e00249da82c96e091b62627499a0b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6079
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>