KUDU-414: Integrate maintenance ops in master
This patch adds a MM in the master and makes sys_catalog pass it to its own tablet. Tested on a
running single node cluster that the operations run.

It could be easy to add the MM to server_base since the code added to the master here is exactly
the same as what's in tablet_server, but it would make the server more Kudu-specific and this
is against our goal of making it reusable in other projects.

Change-Id: I70f5b6394c24574f34d79b03923f66953600e0d3
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6084
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>