KUDU-632 part 1: compaction-test refactors to allow different roll thresholds
Initially, when compaction-test was written, we didn't have
RollingDiskRowSetWriter. Then, when we added the rolling output from
compaction, we didn't update the test to exercise it.

KUDU-632 is a bug which turns up only when the output of a compaction/flush
rolls to multiple output files. This test refactor/cleanup changes around the
APIs a bit so that we can add a new test which actually exercises the rolling
functionality.

Change-Id: Ibdb4f55584c17c12584cab85a62c4da50dbfee6a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6102
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins