KUDU-632 part 2: fix missing UNDO deltas in RollingDiskRowSetWriter output
This fixes a bug where the deltas for the first couple of rows of a DiskRowSet
went missing if that DRS was not the first DRS output by a rolling writer.

The bug was that we had the following sequence:
 - iterate over the input rows, calling AppendDelta() for each, building an
   output block
 - call AppendBlock() for the output block
 -- AppendBlock() decides to roll _before_ writing the block of data.

This results in the deltas being written to the end of the prior rowset rather
than the beginning of the new rowset.

We can't simply change AppendBlock() to roll _after_ appending the block of
data, because then if the amount of data to be flushed exactly filled the
threshold, we'd end up with a rowset with 0 rows, which would probably cause a
lot of assertions to fail. Instead, I rejiggered the code a bit to add an
explicit check for the rolling threshold at the right time.

This adds a new unit test which reliably failed prior to the bug fix. This also
adds a new assertion in the compaction code which ensures that we always write
out UNDO deltas for each row. This doesn't catch the above at exactly the time
it happens. But, it catches the malformed row the next time the broken RowSet
is compacted. I found it useful to track down this bug much more quickly.

Change-Id: I28f0fa9c6d982282dd94c6147eec6ffc72f65dfb
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6103
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins