KUDU-631: TabletPeer MM ops need to go through the peer
The issue in 631 happenend because the MM ops are bypassing the peer and going directly to the log.
The problem with this is we're not TabletPeer::CheckRunning(). Instead of adding that call in the
MM ops file, I encapsulated all that logic in TabletPeer with some refactoring.

I tested that it works with (was usually failing around the 40th run):
$ ./build/release/client-test --gtest_filter=ClientTest.TestDeleteTable --gtest_repeat=1000

Change-Id: I8e7fcda29a0e133a89f61e36aad9cd568bbec532
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6107
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins