KUDU-374: defer cfile footer access to seek time
This is step 0.001 along the path towards fixing KUDU-374, wherein we make
CFile opens as lazy as possible. It doesn't do much on its own, but fell out
of the time I spent reading through the iterator code.

To understand why this is useful you have to understand the scan path.

First, we construct and initialize a whole mess of nested iterators. For a
given scan, this is a union of rowset iterators (filtered by start/end key),
each of which contain a set of cfile iterators (filtered by projection). A
full tablet scan thus constructs an iterator for every cfile in the tablet.

Next, we continue the scan by advancing the iterators. We go rowset by
rowset, advancing its cfile iterators together. Each advance requires a
seek, and each seek implies loading the cfile into memory (so that we can
use its indices). Therefore, a given scan RPC of a full tablet scan will
load all of the cfiles in some subset of rowsets (the actual number of
cfiles depends on the relative size of the rowsets, the number of bytes to
be returned in the scan request, and the relative size of each column).

The current code accesses cfile footers when constructing iterators, which
means all cfiles must be loaded when beginning a full tablet scan. If we
instead wait until we actually seek into a cfile iterator to access the
footer, in the future we could defer cfile loads from open time to scan
time, thus reducing the number of cfiles loaded at the start of a full
tablet scan to just those needed by the first scan RPC, and amortizing the
overall cfile load cost over all of the scan RPCs.

(apologies for the long-winded description, it was intended more for my
understanding than for yours).

Change-Id: Ic90d69dbad04a0ea87da12ac6b51d64b118a06f2
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6108
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>