Fix MergeIterator to respect SelectionVectors in sub-iterators.
Previously, the MergeIterator selection vector was uninitialized,
leading to random rows being selected. MergeIterState (and thus MergeIterator)
now only return rows that are selected by the sub-iterators.

This behavior is verified by a new unit test,
TestMergeIterator::TestPredicate, which performs a scan with a predicate and
validates that the predicate is respected.

Also pursued a few code cleanups while I was in the area. Moved MergeIterator
hasNext from .h to .cc for clarity. Renamed some variables in MergeIterState
to better describe their purpose as counters, rather than pointers.

Change-Id: I3476dd9450ea184a47cfa04259272ffc15c98ccc
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6114
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: Andrew Wang <andrew.wang@cloudera.com>