Rowset metadata gets corrupted when major compacting everything
Sometimes major delta compactions emit an empty delta file, as we've advanced
the base for all columns. In this case we're accounting for this file in the delta tracker, but
not in the row set metadata, which causes the next major compaction to fail.

This patch fixes the problem by only writing the delta file if needed. It also tries to make the
code more readable and it modifies the existing unit test to cover this case.

Change-Id: I256406ec41bad3b8bc0b446077457ca873c6c27b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6120
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>