KUDU-644. Add a test which reproduces a CHECK failure on consensus replica
This adds a test which starts up a single node which is part of a Raft quorum
and then sends hand-crafted RPCs to it, pretending to be a sequence of leaders.
Each new leader aborts the operations written by the previous leader, and adds
some of its own. This reproduces the race condition described in KUDU-644
reasonably quickly.

This also adds a new check in the consensus replica which verifies that the
preceding OpId actually precedes the operations being sent. In the process of
writing the test, I noticed this oversight and figured I'd add the verification.

The new test is initially disabled until the bug can be fixed.

Change-Id: I2fe9349a63df4233328255a98fe4e68f181aa193
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6122
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins