build: put client build artifacts in a build/<type>-client/ directory
For example, a debug client build will now result in build/debug-client/ where
it used to be build/debug/. In all cases, the build/latest/ directory is still
symlinked to point to the most recently selected build type.

This fix makes it more obvious that the artifacts built with
-DKUDU_EXPORTED_CLIENT=1 are actually different than the artifacts built
without. It also allows you to build several different build types, and then
install the artifacts without worrying that a later build will overwrite an
earlier one. This is useful in a packaging build context.

Change-Id: I035be064d4884e9a2dfe9a4cfb34bb1318d49f2c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6126
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins