New integration test for update-heavy workloads
The new test inserts some data then with one thread updates it and with another scans it. It runs
for about 5 seconds if slow tests aren't allowed but if they are then it will use 1M rows and will
run for 100 seconds.

Once we add major compactions to the MM, we'll have a good benchmark to see how compactions affect
throughput. Ad hoc testing those showed >10x performance gains. For the moment, it's useful to see
just how slow scans become by just updating data.

Another thing we need to do is to add a graph to the benchmarks wiki page.

Change-Id: I6c453586195313aa535be1f038b08fc5510af516
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6129
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Jean-Daniel Cryans <jdcryans@cloudera.com>