KUDU-644. Fix HandleFailure to happen in-line on Prepare thread
This changes the HandleFailure logic to happen in-line on the prepare thread in
consensus replicas, rather than deferred to the Apply thread.  This fixes the
potential issue where an operation would be aborted too late, and a later
operation would try to prepare with the same timestamp before the old operation
had a chance to abort.

This re-enables the disabled test, and cleans it up a bit to add assertions
that the correct operations committed, and make it run shorter in quick test
mode.

Change-Id: I6786fa4eb3d4f2f5d25a6421f8ec667145caa4e9
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6130
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins