KUDU-650. Fix stack overflow in RPC connection negotiation when many fds are in use
This switches from using the select() API to poll() instead. select() has the
issue that it can only be used with fds <= 1024. Higher fd values cause a stack
overflow, which was causing a segv on one of our test clusters.

The new unit test crashes in ASAN without the fix. It requires a high ulimit to
run (otherwise it skips itself), so I verified it locally.

Change-Id: I5222ae168d4590d2315177550878fd37b8791305
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6135
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>