codegen: fix NULL termination of embedded IR code
This hopefully will fix the failure of codegen-test in the RELEASE build. It
appears that, even though the LLVM API to parse IR takes a string buffer with
an explicit length, it still expects the string to be NULL-terminated. This
patch changes the build process to NULL-terminate the IR before embedding it in
the file, and also adds a new DCHECK that this is the case.

Change-Id: I86f8a47e334652e712c7a4f7ad1bd13bf29018ca
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6141
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins