KUDU-645: canonicalize and sanitize FsManager paths
What a rathole this turned out to be.

We're now fully canonicalizing the paths provided to FsManager. Because
realpath(3) complains when the last component isn't found, we need to use
DirName() and BaseName() to remove then readd it (which necessitated more
comprehensive versions of those functions).

I didn't like GetAllFilesystemRoots() so I reimplemented it as a proper
Init(), separating constructor input from canonicalized state, the latter of
which needs I/O to be built.

Change-Id: I3140b168ad4e02914de77e964d17031a694c59ed
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6144
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>