KUDU-639 - Enforce term/index ordering for appended/committed operations
This adds a check to make sure that we follow two basic rules when
appending/committing operations:
- Term is monotonically increasing.
- Index is sequential.

Firstly we make sure the leader's request has a proper sequence
of operations before we change anything internally, returning corruption
if something is wrong. This well help pinpoint problems on the leader
side.

Secondly we make sure that any committed operations also follow
a correct sequence, this will help pinpoint problems on the replica
side.

Change-Id: I4d79bf9ac85877d741c57a36205b265b1acbad7d
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6150
Tested-by: David Alves <david.alves@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>