docs: update README to AsciiDoc syntax, fix some inaccuracies
According to Misty, asciidoc is the way to go for docs. This changes
our README file to be AsciiDoc syntax. Along the way, I noticed a few
things that were no longer accurate and improved the content a bit.

It's likely I missed a couple places where the formatting could be better,
but I think docs are an area where any incremental improvement is a good
thing.

Change-Id: Ied1ce309ab85c4dc5dbb0a2cc29843d0f61a4ae2
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6147
Tested-by: jenkins
Tested-by: Misty Stanley-Jones <mstanleyjones@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>