Clarify ownership of LogAnchorRegistry
The TabletPeer should consider itself an owner of the LogAnchorRegistry.
Today, the registry is being kept alive by the MinLogIndexAnchorer in
each memstore, and while that seems to be sufficient, it's confusing and
was an oversight. Now the TabletPeer is explicitly included in the shared
ownership of the registry and the accessor is moved from Tablet to
TabletPeer.

Change-Id: Ib65eb252bafa9376516b36c8cc94f095f82a21fb
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6149
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>