Add support to bind external minicluster daemons to other localhost addresses
This adds a flag (with the default on) to ExternalMiniCluster which allows the
tablet servers to be bound to localhost addresses other than 127.0.0.1.  We
auto-generate a localhost IP within the 127.0.0.1/8 IP block based on the test
process's PID and the index of the daemon within the cluster. On all of the
Linux boxes I checked, it's possible to bind to any address in this range.

It also adds a flag to the Socket class that forces outgoing connections to
bind to a specific IP address. We pass this into the external minicluster
daemons as well, so that outbound connections go from its IP.

With the two of these together, we'll be able to add iptables rules that drop
or reject packets between two hosts, simulating network partitions from within
a single-node test.

A happy side-effect of this patch is that we might see fewer issues with
port collisions from zombie processes causing flakiness on build machines,
because two processes can bind to the same port on different local hosts.

Change-Id: I01fab3a78766514b454722e52713a3c5156d59a7
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6153
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>