KUDU-651 - Consensus: initialize the last received and last committed ids before start
In linked_list-test we were getting a failure because sometimes ConsensusServiceImpl::GetLastOpId()
was returning an uninitialized OpId, which has term and index as mandatory fields, and couldn't
thus be serialized. This was ultimately because RaftConsensusState only initializes
'last_received_opid_' on Start() and sometimes ConsensusServiceImpl::GetLastOpId() would get called
before that.
This changes RaftConsensusState to initialize 'last_received_opid_, as well as
'last_committed_index_', in the constructor so that they are never returned uninitialized.

Change-Id: I24afce6cf965c4390b66d1573e4af8ad1065f9a6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6162
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>