rpc: Move RPC service name from connection setup to per-request
This allows us to pool connections between services, resulting in less
connections required for, say, a service that is both participating in
consensus and also remote bootstrapping.

One security-related downside is that you lose the ability to reject a
connection based on the user not having permissions for that service,
but it's unclear how useful that is, assuming the user passes an
authentication check. The majority of authorization use cases are much
finer grained than "this user is allowed to call this RPC service".

Change-Id: Iec2950be35f8da45c29393c14d6c87a140278ee3
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6164
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>