rpc: Decouple RemoteBootstrapService from TabletService
This simply addresses some hacky stuff that was done prior to KUDU-256.
Prior to that, we had no way share a single port across multiple RPC
services. Now that that capability is there, we should use it for
RemoteBootstrap.

This work was done for some of the other services as a follow-up to
KUDU-256, but this one was not fixed at that time.

Change-Id: I326352d4aa9b5e5f3fdd322fc170a2004e2d9de4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6165
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>