Add a StatusClosure callback
We already have a StatusCallback that is meant to handle the Status result
of some computation but we're missing the converse: an computation
that returns a Status result.

This adds a StatusClosure typedef so that we can pass Closure<Status(void)>
callbacks easily. This also adds a DoNothingStatusClosure() method that
we can bind to by default.

Change-Id: I1911a3697e8829eecf75a90966fd82691d880737
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6167
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>