KUDU-120 - Pre-req. 1 - Make TabletMetadata::Flush() always flush the Log queue
For recovery purposes, we can't ever have a durable TabletMetadata that
points to data blocks that contain data whose corresponding commit Log entries
might still be in flight. In other words, we must always make sure that the Log
is ahead of the persisted tablet data.

The reason for this is pretty clear, if we lose some tablet data that we
were in the process of writing before a crash we can rebuild it from the
Log but the reverse is not possible. That is, were we to write to disk some
batch data whose commit message was not yet in the Log, we wouldn't know,
on replay, whether to reapply that batch or not.

This patch enforces this by adding a 'pre_flush_callback_' to TabletMetadata
which gets called anytime TabletMetadata::Flush() is called successfully.
TabletBootstrap then sets this callback to be Log::WaitForAllFlushed()
effectively enforcing the pre-condition that we want.

Making this a callback instead of a direct call to the Log avoids us having
to have a Log whenever a TabletMetadata is needed (e.g. in a lot of tests).
Having this be an abstract callback can also make our lives easier if we're
to add more pre-conditions later.

Change-Id: I78fe62c30586551574b92f1ba70dcfa303211308
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6171
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins