Make Log extend RefCountedThreadSafe and use a scoped_refptr instead of a raw pointer
The Log is shared across many entities in the code base and its lifetime
should at least be as long as  any of the entities that depend on it. For instance
in some cases Tablet outlives TabletPeer, but the Log doesn't. This will be dangerous
when we make the Tablet flush the Log's entries on Tablet::DoCompactionOrFlush()
(see KUDU-120). We might also want to have scenarios (consensus witnesses) that
don't have a Tablet at all (so maybe they wouldn't have a TabletPeer either).

This patch makes Log extend RefCountedThreadSafe and uses a scoped_refptr for
it where needed.

Change-Id: I4bf5e18f44e9122065c80b99dfd77fc9e09d7dec
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6169
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>