KUDU-642. Several bug fixes for row operation decoding
This expands the fuzz testing for row operation decoding to generate different
pairs of client/server schemas. This exercises various parts of the decoding
code-path, such as dealing with mismatched nullness, mismatched types, invalid
bitmaps, etc.

With ASAN on, and with Arena poisoning enabled (follow-on patch), this exposes
several bugs, one of which was reported as KUDU-642. The patch also fixes those
bugs.

Change-Id: I5a636cad4d91465c681c97c2857607f05693815b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6187
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>