arena: use ASAN manual poisoning to check for Arena allocation overflows
This uses the manual poisoning API [1] to poison the memory within each
Arena component when it is allocated. As we then allocate portions of the
arena to callers, we un-poison the memory, and re-poison it on a Reset.
This catches cases where a caller overruns their allocated buffer within
an Arena.

This also changes the build such that the precompiled IR is built with
-DADDRESS_SANITIZER. This makes the IR code generate calls to the ASAN
manual poisoning APIs even though it is not itself instrumented with ASAN.
That's OK, because the poisoning APIs work by linkage rather than by
instrumentation. Without this change, the generated code crashes whenever
it allocates something from an Arena, because it wouldn't un-poison the
Arena allocation.

I considered enabling ASAN wholesale within the JITted code, but that would
require that the LLVM library we use matches up with the compiler that we are
using for the rest of Kudu. That would prevent us from upgrading to more modern
compilers until we can upgrade our usage of the LLVM libraries. The solution
used here is not quite as nice as fully instrumented JIT code, but it has far
less negative impact on our build flexibility.

Another unexpected complexity of this patch is that, through some extra
transitive includes, precompiled.cc now ends up including the iostream header.
This relies on C++ static destructors, which results in a dependency on the
'__dso_handle' symbol. MCJIT has a bug [2] which requires a workaround in
our JIT module, which I added here in this patch.

[1] https://code.google.com/p/address-sanitizer/wiki/ManualPoisoning
[2] https://llvm.org/bugs/show_bug.cgi?id=18062

Change-Id: I8838adf7303e5f7eb5636b3a58799f0440fb1674
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6186
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins