codegen: fix target dependency calculation for LLVM IR when using Ninja
CMake apparently doesn't support IMPLICIT_DEPENDS with the Ninja generator.
This meant that, when using Ninja, we would often not re-build precompiled.ll
even if header files that it depended on changed.

This patch works around the issue by adding a library containing the
precompiled code, which is built like a normal part of the Kudu source. Nothing
links against this library, but the custom target which generates IR is given
an explicit dependency on it. Thus, the IR code generation ends up inheriting
the correct dependencies.

I tested by touching a dependency header within a fully built tree:

before patch:
  todd@todd-ThinkPad-T540p:~/git/kudu$ ninja -d explain codegen_ir
  ninja: no work to do.
  todd@todd-ThinkPad-T540p:~/git/kudu$ touch src/kudu/util/bitmap.h
  todd@todd-ThinkPad-T540p:~/git/kudu$ ninja -d explain codegen_ir
  ninja: no work to do.

with patch:
  todd@todd-ThinkPad-T540p:~/git/kudu$ ninja -d explain codegen_ir
  ninja: no work to do.
  todd@todd-ThinkPad-T540p:~/git/kudu$ touch src/kudu/util/bitmap.h
  todd@todd-ThinkPad-T540p:~/git/kudu$ ninja -d explain codegen_ir
  ninja explain: output src/kudu/codegen/CMakeFiles/ir_fake_target.dir/precompiled.cc.o older than most recent input src/kudu/util/bitmap.h (1426721482 vs 1426721509)
  ninja explain: src/kudu/codegen/CMakeFiles/ir_fake_target.dir/precompiled.cc.o is dirty
  ninja explain: build/fastdebug/libir_fake_target.a is dirty
  ninja explain: build/fastdebug/precompiled.ll is dirty
  ninja explain: build/fastdebug/precompiled.ll.o is dirty
  ninja explain: build/fastdebug/libcodegen_ir.a is dirty
  [5/5] Linking C static library build/fastdebug/libcodegen_ir.a

Change-Id: Ied60b17fdd8a3c9d271c4b4c9848349e9e063cb7
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6188
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>