build: allow dynamic linking for ASAN
This makes the ASAN build substantially smaller, and thus faster. Whatever bug
we used to hit on older llvm is apparently no longer a problem.

However, if we are using clang 3.4, we can't allow dynamic linking when both
ASAN and UBSAN are enabled. cmake now prevents such a setup and defaults back
to static linking in that case.

Change-Id: Ide0c145b56f32aef10b128f187fecdafdb6e9127
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6189
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>