KUDU-120 - Pre-conditions test
This adds a test that makes sure the preconditions for KUDU-120
are met. That is it makes sure that a compaction is not allowed
to complete, flushing its metadata, before all in flight
transactions whose data may possibly have leaked to the compacted
row sets are both committed and their commit messages are durable
(i.e. have been fsynced to the Log).
This test is supposed to pass both before and after KUDU-120 but
I confirmed manually that it fails when we make a transaction
return to the client right after apply completes without actually
having met the pre-conditions.

Change-Id: I0a325cf01414326806468bfe04c89f56244a34d6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6192
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: David Alves <david.alves@cloudera.com>