KUDU-120 - Part 1 - Commit the mvcc txn right after Apply() and reply to the client
With the pre-requisites in place we can now commit the mvcc transaction and
reply to the client right after we Apply(), instead of having to wait for
commit message to be durable. In fact we can disregard the commit callback
altogether and just finish the transaction right after the commit msg is
enqueued.

This patch gets rid of Transaction::PostCommit() and moves the Finalize()
call on the driver to happen right after we append to the log.

Change-Id: Ia8173736211f02b90aca66307451a50ef05a6721
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6201
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>